requirements.txt
fastapi
uvicorn
python-dotenv
openai
pdf2image
pillow

python-multipart
